refactor(client): flatten transaction lock and commit callback package trees
- Fold client/transaction/lock/models (6 storage-lock classes) and
  lock/metrics (HoodieLockMetrics) into client.transaction.lock, beside
  StorageBasedLockProvider and StorageLockClient they serve; 'models' was a
  filler name. The audit/ subpackage stays (cohesive feature, public API).
- Fold callback/util (HoodieCommitCallbackFactory, HoodieWriteCommitCallbackUtil)
  into org.apache.hudi.callback next to the interfaces they build.
- Move callback/client/http/HoodieWriteCommitHttpCallbackClient to callback.impl
  next to HoodieWriteCommitHttpCallback, its only caller.

FQN-frozen classes (lock providers, callback interfaces/impl/message) untouched.
Behavior-preserving; no moved class is @PublicAPIClass, config-FQN-referenced,
or Kryo-registered.
