+++ hudi-client/hudi-client-common/src/main/java/org/apache/hudi/client/transaction/lock/LockGetResult.java
@@ -16,22 +16,20 @@
  * limitations under the License.
  */
 
-package org.apache.hudi.client.transaction.lock.models;
+package org.apache.hudi.client.transaction.lock;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @AllArgsConstructor
 @Getter
-public enum LockUpsertResult {
-  // Lock was successfully created/updated with code 0
-  SUCCESS(0),
-  // Another process has modified the lock file (precondition failure) with code 1
-  ACQUIRED_BY_OTHERS(1),
+public enum LockGetResult {
+  // Lock file does not exist with code 0
+  NOT_EXISTS(0),
+  // Successfully retrieved the lock file with code 1
+  SUCCESS(1),
   // Unable to determine lock state due to transient errors with code 2
-  UNKNOWN_ERROR(2),
-  // Request was throttled by the storage backend (e.g. HTTP 429) with code 3
-  THROTTLED(3);
+  UNKNOWN_ERROR(2);
 
   private final int code;
 }
