+++ hudi-client/hudi-client-common/src/main/java/org/apache/hudi/client/transaction/lock/StorageBasedLockProvider.java
@@ -21,13 +21,6 @@ package org.apache.hudi.client.transaction.lock;
 import org.apache.hudi.client.transaction.lock.audit.AuditOperationState;
 import org.apache.hudi.client.transaction.lock.audit.AuditService;
 import org.apache.hudi.client.transaction.lock.audit.AuditServiceFactory;
-import org.apache.hudi.client.transaction.lock.metrics.HoodieLockMetrics;
-import org.apache.hudi.client.transaction.lock.models.HeartbeatManager;
-import org.apache.hudi.client.transaction.lock.models.LockGetResult;
-import org.apache.hudi.client.transaction.lock.models.LockProviderHeartbeatManager;
-import org.apache.hudi.client.transaction.lock.models.LockUpsertResult;
-import org.apache.hudi.client.transaction.lock.models.StorageLockData;
-import org.apache.hudi.client.transaction.lock.models.StorageLockFile;
 import org.apache.hudi.common.config.LockConfiguration;
 import org.apache.hudi.common.config.TypedProperties;
 import org.apache.hudi.common.lock.LockProvider;
