+++ hudi-utilities/src/test/java/org/apache/hudi/utilities/callback/TestKafkaCallbackProvider.java
@@ -20,7 +20,7 @@ package org.apache.hudi.utilities.callback;
 
 import org.apache.hudi.callback.HoodieWriteCommitCallback;
 import org.apache.hudi.callback.common.HoodieWriteCommitCallbackMessage;
-import org.apache.hudi.callback.util.HoodieCommitCallbackFactory;
+import org.apache.hudi.callback.HoodieCommitCallbackFactory;
 import org.apache.hudi.common.config.TypedProperties;
 import org.apache.hudi.common.model.HoodieWriteStat;
 import org.apache.hudi.config.HoodieWriteCommitCallbackConfig;
